backend/requirements.txt
openai
python-dotenv
flask
flask-cors
PyYAML
